添加用户类
在model中添加了用户类
